requirements.txt
aiohttp
httpx
py-cord
python-dotenv
aiomysql
